test_login_page now has all tests, deleted tests.py, following pep8
